+++ extension/src/javascripts/content.js
@@ -1,17 +1,11 @@
-var ExtensionId = "lidggcihkifoejcamppcminfbmnnnnjf";
 var CurrentTheme = "default";
-
-function getThemeCss() {
-  var themes = {
-    "okaidia": "assets/themes/okaidia.css",
-    "tomorrow": "assets/themes/tomorrow.css",
-    "coy": "assets/themes/coy.css",
-    "funky": "assets/themes/funky.css",
-    "twilight": "assets/themes/twilight.css",
-    "dark": "assets/themes/dark.css"
-  }
-
-  return themes[CurrentTheme] || null;
+var Themes = {
+  "okaidia": "assets/themes/okaidia.css",
+  "tomorrow": "assets/themes/tomorrow.css",
+  "coy": "assets/themes/coy.css",
+  "funky": "assets/themes/funky.css",
+  "twilight": "assets/themes/twilight.css",
+  "dark": "assets/themes/dark.css"
 }
 
 function importCss() {
@@ -19,7 +13,7 @@ function importCss() {
     "assets/all.css"
   ]
 
-  var theme = getThemeCss();
+  var theme = Themes[CurrentTheme] || null;
   if (theme !== null) {
     array.push(theme);
   }
@@ -33,87 +27,58 @@ function importCss() {
   }
 }
 
-function twoDigits(number) {
-  var str = number + "";
-  if (str.length === 1) {
-    return "0" + str;
-  }
-
-  return str;
-}
-
-function getTimestamp() {
-  var date = new Date();
-  var month = date.getMonth() + 1;
-  var day = date.getDate();
-  var hour = date.getHours();
-  var min = date.getMinutes();
-  var sec = date.getSeconds();
-
-  return date.getFullYear() + twoDigits(month) + twoDigits(day) + twoDigits(hour) + twoDigits(min) + twoDigits(sec);
-}
-
-function isJsonp(text) {
-  return /^[a-zA-Z0-9_$]+\(/.test(text);
-}
-
-function exportJson(codeText, jsonText) {
-  var script = document.createElement("script") ;
-  var json = jsonText;
-  if (isJsonp(codeText)) {
-    json = codeText.replace(/^[a-zA-Z0-9_$]+\(/, '').replace(/\);?$/, '')
-  }
-
-  script.innerHTML = 'window.json = ' + json + ';';
-  document.head.appendChild(script);
-}
-
-function clearPre(pre) {
+function appendCode(code, pre) {
   pre.innerHTML = null;
   pre.removeAttribute("style");
-}
-
-function appendCode(code, pre) {
-  clearPre(pre);
   pre.appendChild(code);
   Prism.highlightElement(code);
 }
 
-function renderJson(pre) {
+function render(pre, jsonText) {
   importCss();
   var codeText = pre.textContent;
-  var jsonText = jsl.format.formatJson(codeText);
+  var formattedText = jsl.format.formatJson(codeText);
 
   var code = document.createElement("code");
   var header = "// " + getTimestamp() + "\n";
   header += "// " + document.location.href + "\n\n";
 
   code.className = "language-json";
-  code.textContent = header + jsonText;
+  code.textContent = header + formattedText;
   appendCode(code, pre);
 
-  exportJson(codeText, jsonText);
+  // Export
+  var script = document.createElement("script") ;
+  script.innerHTML = 'window.json = ' + jsonText + ';';
+  document.head.appendChild(script);
 }
 
-function ready () {
+function getPreWithSource() {
   var childNodes = document.body.childNodes;
   var pre = childNodes[0];
 
   if (childNodes.length === 1 && pre.tagName === "PRE") {
+    return pre;
+  }
+
+  return null
+}
+
+function ready () {
+  var pre = getPreWithSource();
+  if (pre !== null && pre !== undefined) {
     pre.hidden = true;
     try {
+      chrome.runtime.sendMessage({action: "GET_THEME"}, function(response) {
+        CurrentTheme = response.theme;
+        var jsonText = extractJSON(pre.textContent);
 
-      chrome.runtime.sendMessage(ExtensionId, {action: "GET_THEME"}, function(response) {
         try {
-          CurrentTheme = response.theme;
-          renderJson(pre);
-
-        } finally {
-          pre.hidden = false;
-        }
+          JSON.parse(jsonText);
+          render(pre, jsonText);
+        } catch(e) {}
       });
-
-    } catch(e) {
+    } finally {
       pre.hidden = false;
     }
   }