+++ extension/src/javascripts/all.js
@@ -1,6 +1,3 @@
-//= require prism.js
-//= require prism-json.js
-//= require prism-autolinker.js
 //= require jsl-format.js
 //= require content_util.js
 //= require content.js