+++ extension/src/javascripts/all.js
@@ -1,3 +1,5 @@
+//= require prism.js
+//= require prism-code-folder.js
 //= require jsl-format.js
 //= require content_util.js
 //= require content.js