+++ extension/src/javascripts/event.js
@@ -1,6 +1,7 @@
 //= require prism.js
 //= require prism-json.js
 //= require prism-autolinker.js
+//= require prism-code-folder.js
 //= require util.js
 //= require_self
 
@@ -14,9 +15,13 @@ chrome.runtime.onMessage.addListener(function(request, sender, sendResponse) {
     var code = document.createElement("code");
     code.className = "language-json";
     code.innerHTML = request.html;
-    
+
     pre.innerHTML = code.outerHTML;
     pre.className = "language-json";
+    if (getOptions().awaysFold === "true") {
+      pre.setAttribute("data-aways-fold", true);
+    }
+
     code = pre.childNodes[0];
 
     Prism.highlightElement(code);