+++ Changelog.md
@@ -2,6 +2,8 @@
 
   * Options page
   * Six more themes and the ability to change between them
+  * bugfix: activate only on json/jsonp sources
+  * Better json key parser
 
 # 0.1.0