+++ Changelog.md
@@ -4,7 +4,8 @@
   * Six more themes and the ability to change between them
   * bugfix: activate only on json/jsonp sources
   * Better json key parser
-  * Max JSON size option (default: 1024 kb)
+  * Max JSON size option (default: 200 kb)
+  * Uses event page to process the highlight
 
 # 0.1.0