+++ Changelog.md
@@ -1,4 +1,10 @@
-# 0.2.0
+# 0.3.0 - 2013-09-11
+
+  * Collapsible nodes
+  * Option to activate/deactivate nodes always collapsed (from 2º level)
+  * Option to activate/deactivate the header (timestamp + url)
+
+# 0.2.0 - 2013-09-07
 
   * Options page
   * Six more themes and the ability to change between them
@@ -7,6 +13,6 @@
   * Max JSON size option (default: 200 kb)
   * Uses event page to process the highlight
 
-# 0.1.0
+# 0.1.0 - 2013-09-06
 
   * First release with one theme